pyproject.toml
[tool.poetry]
name = "cs-fhe"
version = "0.1.0"
description = ""
authors = ["nutorbit <nutorbitx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
concrete-python = "^2.8.1"
torch = "2.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
